[#27] refactor: 패키지 변경 및 테스트 네이밍 변경
